Refactors shared mock objects to mocks_ddsutil
