+++ ownership/test_views.py
@@ -4,6 +4,7 @@ from rest_framework import status
 from django.test.testcases import TestCase
 from ownership.views import MISSING_TOKEN_MSG, INVALID_TOKEN_MSG, TOKEN_NOT_FOUND_MSG, REASON_REQUIRED_MSG
 from handover_api.models import Handover, State, DukeDSProject, DukeDSUser
+from switchboard.mocks_ddsutil import setup_mock_handover_details
 from django.contrib.auth.models import User as django_user
 from mock import patch, Mock
 
@@ -26,24 +27,6 @@ def create_handover_get_token():
     handover = create_handover()
     return str(handover.token)
 
-
-class MockDDSUser(object):
-    def __init__(self, full_name, email):
-        self.full_name = full_name
-        self.email = email
-
-
-class MockDDSProject(object):
-    def __init__(self, name):
-        self.name = name
-        self.children = []
-
-
-def setup_mock_handover_details(MockHandoverDetails):
-    x = MockHandoverDetails()
-    x.get_from_user.return_value = MockDDSUser('joe', 'joe@joe.com')
-    x.get_to_user.return_value = MockDDSUser('bob', 'bob@joe.com')
-    x.get_project.return_value = MockDDSProject('project')
 
 class AuthenticatedTestCase(TestCase):
     def setUp(self):
